% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{EStep}
\alias{EStep}
\title{E step of the EM algorithm: Calculate the expectation of Q function given parameters}
\usage{
EStep(u, pi, mu1, mu2, var1, var2)
}
\arguments{
\item{u}{The combined data of (y, x)}

\item{pi}{The proportion of cell type 1 in all cells}

\item{mu1}{The estimated (y,x) mean function of cell type 1}

\item{mu2}{The estimated (y,x) mean function of cell type 2}

\item{var1}{The estimated (y,x) variance of cell type 1}

\item{var2}{The estimated (y,x) variance of cell type 2}
}
\value{
A list of the 4 elements
\item{Eu1:}{The expected value in cell type 1 given the observed (y, x) at the tissue level given input parameters}
\item{Eu2:}{The expected value in cell type 2 given the observed (y, x) at the tissue level given input parameters}
\item{Evar1:}{The expected variance-covariance matrix in cell type 1 given the observed (y, x) at the tissue level given input parameters}
\item{Evar2:}{The expected variance-covariance matrix in cell type 2 given the observed (y, x) at the tissue level given input parameters}
}
\description{
E step of the EM algorithm: Calculate the expectation of Q function given parameters
}
