% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProMix.eFDR.PermReplace.R
\name{iProMix.eFDR.PermReplace}
\alias{iProMix.eFDR.PermReplace}
\title{FDR controlled iProMix identification with permutation-replace procedure}
\usage{
iProMix.eFDR.PermReplace(
  yMatrix,
  x,
  cov = NULL,
  pi,
  reduce1 = c(2, 1),
  reduce2 = NULL,
  cl = F,
  B = 1,
  seed = NULL,
  FDR = 0.1
)
}
\arguments{
\item{yMatrix}{The quantitative measure (e.g. protein/expression) of a gene in the matrix form (Row: Genes; Columns: Samples)}

\item{x}{The quantitative measure of anther gene  (e.g. ACE2 protein levels) that we would like to know their cell-type specific dependency with Y}

\item{cov}{The covariates for adjustment. Their impact on the mean value of X and Y are adjusted}

\item{pi}{The proportion of cell type 1}

\item{reduce1}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 1.  Default is NULL.}

\item{reduce2}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 2.  Default is NULL.}

\item{cl}{True of False. If true, parallel computing is used; need the library(doRNG). Default is FALSE}

\item{B}{The number of permutation}

\item{seed}{Seed for the permutation}

\item{FDR}{FDR cutoff; Default 0.1}
}
\value{
list with 4 elements. It contains
\item{NoSigGene:}{Number of significant genes at the pre-determined FDR cutoff}
\item{IdxSigGene:}{The index of significant genes}
\item{ft_data:}{The iProMix output (e.g. estimated parameters, log likelihoods) using original data}
\item{ft_permutated:}{The iProMix output (e.g. estimated parameters, log likelihoods) using permuted data}
}
\description{
FDR controlled iProMix identification with permutation-replace procedure
}
\examples{
\donttest{library(iProMix)
set.seed(111)
y <- matrix(rnorm(100,10,1), ncol=10)
x <- rnorm(10,10,1)
pi <- runif(10)
result <- iProMix.eFDR.PermReplace(yMatrix=y, x=x, pi=pi,reduce1=c(2,1), B=1, seed=1132, FDR=0.1)
}
}
