% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProMix.LRT.R
\name{iProMix.LRT}
\alias{iProMix.LRT}
\title{Calculate the likelihood ratio test statistics (as well as chisq p-value) for iProMix}
\usage{
iProMix.LRT(
  y,
  x,
  cov = NULL,
  pi,
  reduce1 = c(2, 1),
  reduce2 = NULL,
  tuningPar = 1e-06,
  diffNum = 1e-04,
  numitersNum = 200
)
}
\arguments{
\item{y}{The quantitative measure (e.g. protein/expression) of a gene}

\item{x}{The quantitative measure of anther gene  (e.g. ACE2 protein levels) that we would like to know their cell-type specific dependency with Y}

\item{cov}{The covariates for adjustment. Their impact on the mean value of X and Y are adjusted}

\item{pi}{The proportion of cell type 1}

\item{reduce1}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 1.  Default is NULL.}

\item{reduce2}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 2.  Default is NULL.}

\item{tuningPar}{Default is 1e-8. It is used in the embedded graphic lasso procedure for estimating correlation. A larger tuningPar can be selected if one is interested in penalized estimates.}

\item{diffNum}{Default 0.0001. The convergency criterion for EM algorithm.}

\item{numitersNum}{Default 200. The number of iterations in EM algorithm.}
}
\value{
list with 10 elements. It contains
\item{var1:}{The estimated (y,x) variance of cell type 1}
\item{var2:}{The estimated (y,x) variance of cell type 2}
\item{mu1:}{The estimated (y,x) mean function of cell type 1}
\item{mu2:}{The estimated (y,x) mean function of cell type 2}
\item{cor.score1:}{The estimated X-Y correlation in cell type 1}
\item{cor.score2:}{The estimated X-Y correlation in cell type 2}
\item{full.ll:}{The estimated log likelihood function from the full model}
\item{reduced.ll:}{The estimated log likelihood function from the reduced model}
\item{LRT:}{The likelihood ratio test statistics for iProMix}
\item{LRT.pvalue:}{The chisq p-value for iProMix}
}
\description{
Calculate the likelihood ratio test statistics (as well as chisq p-value) for iProMix
}
\examples{
library(iProMix)
set.seed(111)
y <- rnorm(100,10,1)
x <- rnorm(100,10,1)
pi <- runif(100)
iProMix.LRT_result <- iProMix.LRT(y = y, x = x, pi = pi, reduce1=c(2,1), reduce2=NULL)
}
