% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProMix.PermAdd.LRT.matrix.R
\name{iProMix.PermAdd.LRT.matrix}
\alias{iProMix.PermAdd.LRT.matrix}
\title{For Y matrix, calculate the difference in likelihood ratio test statistics for x and x_tilde}
\usage{
iProMix.PermAdd.LRT.matrix(
  yMatrix,
  x,
  x_tilde = NULL,
  cov = NULL,
  pi,
  CellType = 1,
  B = 1,
  seed = NULL,
  tuningPar = 1e-06,
  cl = F
)
}
\arguments{
\item{yMatrix}{The quantitative measure (e.g. protein/expression) of a gene in the matrix form (Row: Genes; Columns: Samples)}

\item{x}{The quantitative measure of anther gene  (e.g. ACE2 protein levels) that we would like to know their cell-type specific dependency with Y}

\item{x_tilde}{The permuatated x}

\item{cov}{The covariates for adjustment. Their impact on the mean value of X and Y are adjusted}

\item{pi}{The proportion of cell type 1}

\item{CellType}{The cell type to be estimated. Default is 1.}

\item{B}{The number of permutation}

\item{seed}{Seed for the permutation}

\item{tuningPar}{Default is 1e-8. It is used in the embedded graphic lasso procedure for estimating correlation. A larger tuningPar can be selected if one is interested in penalized estimates.}

\item{cl}{True of False. If true, parallel computing is used; need the library(doRNG). Default is FALSE}
}
\value{
list with number of elements equal to the number of genes for the yMatrix. Each gene contains a sublist with 2 elements. It contains
\item{ll:}{The estimated log likelihood function}
\item{M:}{The M value}
}
\description{
For Y matrix, calculate the difference in likelihood ratio test statistics for x and x_tilde
}
\examples{
\donttest{library(iProMix)
set.seed(111)
y <- matrix(rnorm(1000,10,1), ncol=10)
x <- rnorm(100,10,1)
pi <- runif(100)
result <- iProMix.PermAdd.LRT.matrix(yMatrix=y, x=x, pi=pi, B=1, CellType=1)
}
}
