% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProMix.R
\name{iProMix}
\alias{iProMix}
\title{A decomposition model}
\usage{
iProMix(
  y,
  x,
  cov = NULL,
  pi,
  tuningPar = 1e-08,
  diffNum = 1e-04,
  numitersNum = 200,
  reduce1 = NULL,
  reduce2 = NULL,
  inital.mu1 = NULL,
  inital.mu2 = NULL,
  inital.var1 = NULL,
  inital.var2 = NULL
)
}
\arguments{
\item{y}{The quantitative measure (e.g. protein/expression) of a gene}

\item{x}{The quantitative measure of anther gene  (e.g. ACE2 protein levels) that we would like to know their cell-type specific dependency with Y}

\item{cov}{The covariates for adjustment. Their impact on the mean value of X and Y are adjusted}

\item{pi}{The proportion of cell type 1}

\item{tuningPar}{Default is 1e-8. It is used in the embedded graphic lasso procedure for estimating correlation. A larger tuningPar can be selected if one is interested in penalized estimates}

\item{diffNum}{Default 0.0001. The convergency criterion for EM algorithm.}

\item{numitersNum}{Default 200. The number of iterations in EM algorithm.}

\item{reduce1}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 1.  Default is NULL.}

\item{reduce2}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 2.  Default is NULL.}

\item{inital.mu1}{The initial mean value of (Y,X) for cell type 1, starting in Y and then in X. At default inital.mu1 =NULL, tissue-level mean is used.}

\item{inital.mu2}{The initial mean value of (Y,X) for cell type 2, starting in Y and then in X. At default inital.mu2 =NULL, tissue-level mean is used.}

\item{inital.var1}{The initial variance-covariance matrix of (Y,X) for cell type 1. At default inital.var1 =NULL, tissue-level variance-covariance matrix is used.}

\item{inital.var2}{The initial variance-covariance matrix of (Y,X) for cell type 2. At default inital.var2 =NULL, tissue-level variance-covariance matrix is used.}
}
\value{
list with 9 elements. It contains
\item{var1:}{The estimated (y,x) variance of cell type 1}
\item{var2:}{The estimated (y,x) variance of cell type 2}
\item{mu1:}{The estimated (y,x) mean function of cell type 1}
\item{mu2:}{The estimated (y,x) mean function of cell type 2}
\item{cor.score1:}{The estimated X-Y correlation in cell type 1}
\item{cor.score2:}{The estimated X-Y correlation in cell type 2}
\item{ll:}{The estimated log likelihood function}
\item{coef1:}{The estimated covariate effects on (y,x) in cell type 1}
\item{coef2:}{The estimated covariate effects on (y,x) in cell type 2}
}
\description{
A decomposition model
}
\examples{
library(iProMix)
set.seed(111)
y <- rnorm(100,10,1)
x <- rnorm(100,10,1)
pi <- runif(100)
iProMix_result <- iProMix(y = y, x = x, pi = pi, reduce1=c(2,1), reduce2=NULL)
}
