% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{Calculate the log likelihood function given parameters}
\usage{
LL(u, pi, mu1, mu2, var1, var2)
}
\arguments{
\item{u}{The combined data of (y, x)}

\item{pi}{The proportion of cell type 1 in all cells}

\item{mu1}{The estimated (y,x) mean function of cell type 1}

\item{mu2}{The estimated (y,x) mean function of cell type 2}

\item{var1}{The estimated (y,x) variance of cell type 1}

\item{var2}{The estimated (y,x) variance of cell type 2}
}
\value{
log likelihood function
}
\description{
Calculate the log likelihood function given parameters
}
