% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{MStep}
\alias{MStep}
\title{M step of the EM algorithm: Update the parameter estimates given the expectation of Q functions}
\usage{
MStep(Eu1, Eu2, Evar1, Evar2, cov, reduce1, reduce2, tuningPar = 1e-08)
}
\arguments{
\item{Eu1}{The expected value in cell type 1 given the observed (y, x) at the tissue level}

\item{Eu2}{The expected value in cell type 2 given the observed (y, x) at the tissue level}

\item{Evar1}{The expected variance-covariance matrix in cell type 1 given the observed (y, x) at the tissue level}

\item{Evar2}{The expected variance-covariance matrix in cell type 2 given the observed (y, x) at the tissue level}

\item{cov}{The covariates under consideration for its effects on the mean function of (y, x)}

\item{reduce1}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 1.  Default is NULL.}

\item{reduce2}{A index of the row and column of the variance-covariance matrix that should be forced to be zero in cell type 2.  Default is NULL.}

\item{tuningPar}{Default is 1e-8. It is used in the embedded graphic lasso procedure for estimating correlation. A larger tuningPar can be selected if one is interested in penalized estimates.}
}
\value{
A list of the 5 elements
\item{var1:}{The estimated (y,x) variance of cell type 1}
\item{var2:}{The estimated (y,x) variance of cell type 1}
\item{mu1:}{The estimated (y,x) mean function of cell type 1}
\item{mu2:}{The estimated (y,x) mean function of cell type 2}
\item{coef1:}{The estimated covariate effects on (y,x) in cell type 1}
\item{coef2:}{The estimated covariate effects on (y,x) in cell type 2}
}
\description{
M step of the EM algorithm: Update the parameter estimates given the expectation of Q functions
}
